% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_gen.R
\name{app_data_gen}
\alias{app_data_gen}
\title{app_data_gen: Generate Example Data for use with EligModelShinyApp()}
\usage{
app_data_gen(n = 1000, n_trials = 3, n_rwd = 2, n_cohorts = 1, miss_pct = 0.1)
}
\arguments{
\item{n}{(type = integer) number of observations (patients) to include in the example data (default=1000)}

\item{n_trials}{(type = integer) number of clinical trials in the example data (default=3)}

\item{n_rwd}{(type = integer)  number of real-world datasets in the example data (default=2)}

\item{n_cohorts}{type = integer) number of patient cohorts in the example data (default=1)}

\item{miss_pct}{(type = numeric) proportion missing in the example eligibility criteria data (default=0.10)}
}
\value{
A named list with the following components:
\itemize{
\item elig_dat_list - list of data.frames with example eligibility criteria data. The length of the list equals the number of trial/RWD combinations specified.
\item char_dat_list - list of data.frames with example patient characteristic data. The length of the list equals the number of trial/RWD combinations specified.
\item surv_dat_list - list of data.frames with example survival data. The length of the list equals the number of trial/RWD combinations specified.
\item data_dict_list - list of data.frames with example data dictionaries. The length of the list equals the number of trial/RWD combinations specified.
}
}
\description{
Generates a list of example datasets to use as input for EligModelShinyApp(). The returned list includes
an example list for patient eligibility data (elig_dat_list), patient characteristics (char_dat_list),
time-to-event data (surv_dat_list), and data dictionaries (data_dict_list). The length of each list
corresponds to the number of trial and real-world dataset combinations specified.
}
