% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elig_data_gen.R
\name{elig_data_gen}
\alias{elig_data_gen}
\title{elig_data_gen: Generate example eligibility data}
\usage{
elig_data_gen(n = 1000, dim_X = 10, rho = 0.3, miss_pct = 0.1)
}
\arguments{
\item{n}{Number of observations (default=1000)}

\item{dim_X}{Number of baseline covariates (default=10)}

\item{rho}{Pairwise correlation of X variables (default=0.30)}

\item{miss_pct}{Missingness percentage (assumes MCAR, default=0.10)}
}
\value{
\itemize{
\item Y (continuous outcome of interest)
\item X (baseline covariates, with missing)
\item elig_dat (matrix of eligibility indicators; for X1-X3)
}
}
\description{
Function that generates example eligibility data. Key outputs include Y (outcome of interest),
and X (baseline covariates; includes eligibility variables). Participants are considered eligible
if X1>0, X2>0, and X3>0.
}
