% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elig_model.R
\name{elig_model}
\alias{elig_model}
\title{elig_model: Estimate clinical eligibility probabilities}
\usage{
elig_model(
  Y,
  X,
  Xtest = NULL,
  elig.dat = NULL,
  elig.rules = NULL,
  elig.mod = "filter",
  elig.hyper = NULL,
  elig.CV = TRUE,
  V = 5,
  stratify = TRUE,
  m_imp = 5,
  ...
)
}
\arguments{
\item{Y}{Outcome. Must be numeric for continuous or binary outcomes. For survival outcomes, must be in Surv() form.}

\item{X}{Covariate space. Must be data.frame()}

\item{Xtest}{Test data-set}

\item{elig.dat}{data.frame() with columns corresponding to individual eligibility criteria. Each column must be
binary (0, 1, or NA; which correspond to not-satisfying, satisfying, or uncertain about that specific criteria).
If not provided, must provide "elig.rules" (see below).}

\item{elig.rules}{Data-frame of eligibility rules. For example, data.frame(var = c("X1", "X2), thres=c(">0", ">0")),
which indicates that a patient is eligibile if X1>0 and X2>0. If not provided, must provide "elig.dat".}

\item{elig.mod}{Eligibility model. Default=NULL. Other options include "filter", which only include patients
who satisfy all eligibility criteria; or multivariate outcome random forest ("rfmv")
through randomForestSRC; or multiple imputation via chained equations ("mice_rf"). For all approaches,
elig.dat must  also be supplied. See below for details.}

\item{elig.hyper}{Hyper-parameters for eligibility model, must be list. Default=NULL.}

\item{elig.CV}{Use cross-validation in eligibility model? Default=TRUE.}

\item{V}{Number of cross-validation folds. Default=5, only applicable if elig.CV=T.}

\item{stratify}{For CV, stratify by eligiblity status? Default=TRUE.}

\item{m_imp}{Number of imputed data-sets (only used for elig.mod="mice_rf", default=5)}

\item{...}{Other arguments to pass through}
}
\value{
\itemize{
\item mod - Fitted eligibility model(s)
\item elig.dat - Eligibility output (elig indicator/probabilities/weights)
\item res.CV - cross-validation results (NULL if elig.CV=FALSE)
\item lambda.dat - Trimming threshold results
}
}
\description{
For analyses including RWD and clinical trial data, one approach to make the populations
more similar is to apply the available clinical eligibility criteria on the RWD
patient. For example, only include RWD patients with ECOG<=1 and Age>18. Current options
include elig.mod="filter" (only include patients who satisfy all criteria),
elig.mod="simple" (for each patient, estimate the average number of criteria satisfied;
use this as a weight), and elig.mod="rfmv" (use multivariate random forest through
randomForestSRC to estimate the probability of satisfying all criteria).
}
\examples{

## Testing ##
\donttest{
library(eligmodel)
library(dplyr)
library(survival)

# Generate example data #
dat <- elig_data_gen(n=1000, dim_X=10, rho=0.3, miss_pct=0.1)
Y <- dat$Y
X <- dat$X
elig.dat <- dat$elig_dat

# Test Different Approaches #

# Filtering #
mod1 <- elig_model(Y=Y, X=X, elig.dat=elig.dat, elig.mod="filter")
summary(mod1$elig.dat)
aggregate(w_elig ~ elig, data=mod1$elig.dat, FUN="summary")
plot_elig(mod1)
plot_elig(mod1, type="outcome")
plot_elig(mod1, type="balance") # Differences should be zero, IE=1 is same as filter

# Simple Weighting #
mod2 <- elig_model(Y=Y, X=X, elig.dat=elig.dat, elig.mod="simple")
summary(mod1$elig.dat)
aggregate(w_elig ~ elig, data=mod2$elig.dat, FUN="summary")
plot_elig(mod2)
plot_elig(mod2, type="outcome")
plot_elig(mod2, type="balance")

# Model-based (multivariate RF) ##
mod3 <- elig_model(Y=Y, X=X, elig.dat=elig.dat, elig.mod="rfmv")
summary(mod3$elig.dat)
aggregate(w_elig ~ elig, data=mod3$elig.dat, FUN="summary")
plot_elig(mod3)
plot_elig(mod3, type="outcome")
plot_elig(mod3, type="balance")

# MICE RF #
elig.rules <- data.frame(var = c("X1", "X2", "X3"),
                         thres = c(">= 0", ">0", ">0"))
mod4 <- elig_model(Y=Y, X=X, elig.mod="mice_rf", elig.rules = elig.rules)
summary(mod4$elig.dat)
aggregate(w_elig ~ elig, data=mod4$elig.dat, FUN="summary")
plot_elig(mod4)
plot_elig(mod4, type="outcome")
plot_elig(mod4, type="balance")

}
}
