% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elig.R
\name{plot_elig}
\alias{plot_elig}
\title{plot_elig: Plot diagnostic function for eligibility modeling.}
\usage{
plot_elig(fit, type = "histo")
}
\arguments{
\item{fit}{A fitted \code{rwdcausal} or fitted \code{propensity_model} object.}

\item{type}{Type of plot. Options include "histo" (default: Histogram of eligibility weights by status), "outcome"
(Depending on outcome Y, mean estimates (continuous or binary) or Kaplan-Meier curves (survival) by eligibility status,
and "balance" (calculate standardized mean differences for each covariate between eligible and weighted groups).}
}
\value{
Eligibility diagnostic plot (ggplot2 object)
}
\description{
This function currently plots weights by eligibility status, estimates of Y by eligibility status,
and covariate balance across the X variables between IE=1 vs Weighted populations.
}
